just fixed the error handling of services>deletes

+++ src/main/java/com/dmggg/train_app/services/agroupament/AgroupamentService.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,6 +16,7 @@ import com.dmggg.train_app.entities.exercise.Exercise;
 import com.dmggg.train_app.repositories.agroupament.AgroupamentRepository;
 import com.dmggg.train_app.repositories.agroupament.SubAgroupamentRepository;
 import com.dmggg.train_app.repositories.exercise.ExerciseRepository;
+import com.dmggg.train_app.services.exceptions.DatabaseException;
 import com.dmggg.train_app.services.exceptions.EntityNotFound;
 
 import lombok.RequiredArgsConstructor;
@@ -27,8 +29,6 @@ public class AgroupamentService {
   private final ExerciseRepository exerciseRepository;
   private final AgroupamentRepository repository;
 
-
-
   @Transactional(readOnly = true)
   public List<AgroupamentResponse> searchAll() {
     List<Agroupament> list = repository.findAll();
@@ -37,20 +37,12 @@ public class AgroupamentService {
     return listResponse;
   }
 
-
-
-
-
   @Transactional(readOnly = true)
   public AgroupamentResponse searchById(Long id) {
     Optional<Agroupament> agroupament = repository.findById(id);
-    return new AgroupamentResponse(agroupament.orElseThrow(()
-        -> new EntityNotFound("Agroupament not found on our database")));
+    return new AgroupamentResponse(
+        agroupament.orElseThrow(() -> new EntityNotFound("Agroupament not found on our database")));
   }
-
-
-
-
 
   @Transactional
   public AgroupamentResponse insert(AgroupamentRequest agroupamentRequest) {
@@ -61,7 +53,7 @@ public class AgroupamentService {
     if (agroupamentRequest.getExercises() != null && !agroupamentRequest.getExercises().isEmpty()) {
       agroupamentRequest.getExercises().forEach(x -> {
         Exercise exercise = exerciseRepository.findById(x)
-        .orElseThrow(() -> new EntityNotFound("Exercise id :" + x + " not found on our database"));
+            .orElseThrow(() -> new EntityNotFound("Exercise id :" + x + " not found on our database"));
 
         exercise.getListAgroupaments().add(agroupament);
         exerciseRepository.save(exercise);
@@ -71,7 +63,7 @@ public class AgroupamentService {
     if (agroupamentRequest.getSubAgroupaments() != null && !agroupamentRequest.getSubAgroupaments().isEmpty()) {
       agroupamentRequest.getSubAgroupaments().forEach(x -> {
         SubAgroupament subAgroupament = subAgroupamentRepository.findById(x)
-        .orElseThrow(() -> new EntityNotFound("SubAgroupament id :" + x + " not found on our database"));
+            .orElseThrow(() -> new EntityNotFound("SubAgroupament id :" + x + " not found on our database"));
 
         subAgroupament.setAgroupament(agroupament);
         subAgroupamentRepository.save(subAgroupament);
@@ -79,13 +71,9 @@ public class AgroupamentService {
     }
 
     Agroupament inserted = repository.save(agroupament);
-    
+
     return new AgroupamentResponse(inserted);
   }
-
-
-
-
 
   @Transactional
   public AgroupamentResponse update(Long id, AgroupamentRequest agroupamentRequest) {
@@ -95,19 +83,16 @@ public class AgroupamentService {
     agroupament.setName(agroupamentRequest.getName());
 
     if (agroupamentRequest.getSubAgroupaments() != null) {
-      new ArrayList<>(agroupament.getListSubAgroupaments()).forEach(x ->
-        {
-          if (!agroupamentRequest.getSubAgroupaments().contains(x.getId())) {
-            x.setAgroupament(null);
-            subAgroupamentRepository.save(x);
-          }
+      new ArrayList<>(agroupament.getListSubAgroupaments()).forEach(x -> {
+        if (!agroupamentRequest.getSubAgroupaments().contains(x.getId())) {
+          x.setAgroupament(null);
+          subAgroupamentRepository.save(x);
         }
-      );
-      
+      });
 
       agroupamentRequest.getSubAgroupaments().forEach(x -> {
         SubAgroupament subAgroupament = subAgroupamentRepository.findById(x)
-        .orElseThrow(() -> new EntityNotFound("SubAgroupament id :" + x + " not found on our database"));
+            .orElseThrow(() -> new EntityNotFound("SubAgroupament id :" + x + " not found on our database"));
 
         if (!agroupament.getListSubAgroupaments().contains(subAgroupament)) {
           subAgroupament.setAgroupament(agroupament);
@@ -117,21 +102,17 @@ public class AgroupamentService {
       });
     }
 
-
     if (agroupamentRequest.getExercises() != null) {
-      new ArrayList<>(agroupament.getListExercises()).forEach(x ->
-        {
-          if (!agroupamentRequest.getExercises().contains(x.getId())) {
-            x.getListAgroupaments().remove(agroupament);
-            exerciseRepository.save(x);
-          }
+      new ArrayList<>(agroupament.getListExercises()).forEach(x -> {
+        if (!agroupamentRequest.getExercises().contains(x.getId())) {
+          x.getListAgroupaments().remove(agroupament);
+          exerciseRepository.save(x);
         }
-      );
-      
+      });
 
       agroupamentRequest.getExercises().forEach(x -> {
         Exercise exercise = exerciseRepository.findById(x)
-        .orElseThrow(() -> new EntityNotFound("Exercise id :" + x + " not found on our database"));
+            .orElseThrow(() -> new EntityNotFound("Exercise id :" + x + " not found on our database"));
 
         if (!agroupament.getListExercises().contains(exercise)) {
           exercise.getListAgroupaments().add(agroupament);
@@ -145,25 +126,31 @@ public class AgroupamentService {
     return new AgroupamentResponse(updated);
   }
 
-
-
-
-
   @Transactional
   public void delete(Long id) {
-    Agroupament agroupament = repository.findById(id)
-        .orElseThrow(() -> new EntityNotFound("Agroupament not found on our database"));
-    
-    new ArrayList<>(agroupament.getListSubAgroupaments()).forEach(x -> {
-      x.setAgroupament(null);
-      subAgroupamentRepository.save(x);
-    });
+    if (!repository.existsById(id)) {
+      throw new EntityNotFound("Agroupament not found on our database");
+    }
+    try {
 
-    new ArrayList<>(agroupament.getListExercises()).forEach(x -> {
-      x.getListAgroupaments().remove(agroupament);
-      exerciseRepository.save(x);
-    });
+      Agroupament agroupament = repository.findById(id)
+          .orElseThrow(() -> new EntityNotFound("Agroupament not found on our database"));
 
-    repository.delete(agroupament);
+      new ArrayList<>(agroupament.getListSubAgroupaments()).forEach(x -> {
+        x.setAgroupament(null);
+        subAgroupamentRepository.save(x);
+      });
+
+      new ArrayList<>(agroupament.getListExercises()).forEach(x -> {
+        x.getListAgroupaments().remove(agroupament);
+        exerciseRepository.save(x);
+      });
+
+      repository.delete(agroupament);
+
+    } catch (DataIntegrityViolationException e) {
+      throw new DatabaseException("Integrity violation");
+    }
+
   }
 }
